refactor: 下沉 Agent 权限与记忆契约
