+++ packages/agentcore/lib/agentcore.dart
@@ -3,3 +3,5 @@ library;
 export 'src/agent_core.dart';
 export 'src/contracts.dart';
 export 'src/json_turn_parser.dart';
+export 'src/memory/agent_memory.dart';
+export 'src/permission/agent_authorization.dart';
